backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
dashscope>=1.25.0
python-dotenv==1.0.0
werkzeug==3.0.1
tiktoken==0.5.2
